feat: create new my courses tab in courses page
